% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaperm-package.R
\docType{package}
\name{adaperm-package}
\alias{adaperm-package}
\title{Permutation tests for adaptive designs

Permutation and re-randomization tests for adaptive designs.

Provides functions to perform permutation tests.}
\description{
Permutation tests for adaptive designs

Permutation and re-randomization tests for adaptive designs.

Provides functions to perform permutation tests.
}

