% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaperm}
\alias{adaperm}
\title{Permutation based test for adaptive designs}
\usage{
adaperm(x, g = NULL, n1, n, m1 = n1, m = n, test_statistic,
  alpha0 = NULL, alpha = 0.025, cer_type = "non-randomized",
  atest_type = "non-randomized", permutations = 10000, stratified = TRUE)
}
\arguments{
\item{x}{Observations}

\item{g}{Treatment assignments (if \code{NULL}) a one-sample test will be performed}

\item{n1}{First stage control-group sample size}

\item{n}{Pre-planned overall sample size}

\item{m1}{First stage treatment-group sample size (will be ignored if \code{g} is \code{NULL})}

\item{m}{Pre-planned overall treatment-group sample size (will be ignored if \code{g} is \code{NULL})}

\item{test_statistic}{Test statistic}

\item{alpha0}{Early rejection boundary for group sequential trial}

\item{alpha}{Significance level}

\item{cer_type}{what type of conditional error rate function should be used (see \code{\link{permutation_cer}}) for detailcer_type}

\item{atest_type}{if 'CER' compute only conditional error rate, else type of adaptive test should be performed (see \code{\link{perm_test}} for details)}

\item{permutations}{Number of permutations to use}

\item{stratified}{should permutation be stratified by stage}
}
\value{
An object of class \code{htest}
}
\description{
Main user interface for adaperm
}
\details{
Group sequentialism assumes that the trial does not stop at the first stage - otherwise second stage data would not be available.
}
\author{
Florian Klinglmueller
}

