% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaptive_DR_os}
\alias{adaptive_DR_os}
\title{One-sample adaptive permutation test}
\usage{
adaptive_DR_os(x, n1, n, ne, test_statistic, perms = 50000, alpha = 0.025)
}
\arguments{
\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{perms}{Maximum number of permutations to use when computing permutation p-values and conditional error rates}
}
\value{
p-value
}
\description{
Adaptive decision rule
}
\details{
Slightly faster than the test that computes p-values
}
\author{
Florian Klinglmueller
}

