% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{adaptive_invnormtest_2s}
\alias{adaptive_invnormtest_2s}
\title{Inverse normal adaptive wilcoxon test}
\usage{
adaptive_invnormtest_2s(x, y, n1, n, ne, m1 = n1, m = n, me = ne,
  alpha = 0.025)
}
\arguments{
\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{alpha}{}
}
\value{
Logical whether the test rejects
}
\description{
Adaptive combination test of stage-wise 2 samples t-tests using the inverse normal combination function.
}
\author{
Florian Klinglmueller
}

