% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaptive_permdr}
\alias{adaptive_permdr}
\title{Quick adaptive permutation test}
\usage{
adaptive_permdr(x1, x2, xE, g1, g2, gE, test_statistic, alpha, permutations,
  restricted, atest_type = c("non-randomized", "mid-p", "davison_hinkley",
  "CER"), cer_type = c("non-randomized", "randomized", "uniform"))
}
\arguments{
\item{x1}{first stage observations}

\item{x2}{second stage observations}

\item{xE}{extended stage observations}

\item{g1}{frist stage treatment assignments}

\item{g2}{second stage treatment assignments}

\item{gE}{extended stage treatment assignments}

\item{test_statistic}{test statistic to use}

\item{alpha}{significance level}

\item{permutations}{number of permutations to use}

\item{restricted}{should the treatment group sizes be fixed}

\item{atest_type}{if 'CER' compute only conditional error rate, else type of adaptive test should be performed (see \code{\link{perm_test}} for details)}

\item{cer_type}{what type of conditional error rate function should be used (see \code{\link{permutation_cer}}) for details}
}
\value{
pvalue
}
\description{
Quick no-nonsens adaptive permutation decision rule
}
\author{
Florian Klinglmueller
}

