% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaptive_permtest2_os}
\alias{adaptive_permtest2_os}
\title{One-sample adaptive permutation test with subsampling}
\usage{
adaptive_permtest2_os(x, n1, n, ne, test_statistic, perms = 50000,
  resam = 100, alpha = 0.025)
}
\arguments{
\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{perms}{Maximum number of permutations to use when computing permutation p-values and conditional error rates}

\item{combination_function}{Function to combine stage-wise (permutation) p-values}
}
\value{
p-value
}
\description{
Adaptive permutation test one-sample problems
}
\author{
Florian Klinglmueller
}

