% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaptive_permtest_quick}
\alias{adaptive_permtest_quick}
\title{Quick adaptive permutation test}
\usage{
adaptive_permtest_quick(x1, x2, xE, g1, g2, gE, stat, permutations = 10000,
  restricted = T, stratified = T)
}
\arguments{
\item{x1}{first stage observations}

\item{x2}{second stage observations}

\item{xE}{extended stage observations}

\item{g1}{frist stage treatment assignments}

\item{g2}{second stage treatment assignments}

\item{gE}{extended stage treatment assignments}

\item{stat}{test statistic to use}

\item{permutations}{number of permutations to use}

\item{restricted}{should the treatment group sizes be fixed}
}
\value{
pvalue
}
\description{
Quick no-nonsens adaptive permutation test that computes only the p-value
}
\author{
Florian Klinglmueller
}

