% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaptive_permtest_ts}
\alias{adaptive_permtest_ts}
\title{Two-sample adaptive permutation test}
\usage{
adaptive_permtest_ts(x, y, n1, n, ne, m1 = n1, m = m, me = me,
  test_statistic, perms = 50000, alpha = 0.025)
}
\arguments{
\item{x}{Observations control group}

\item{y}{Observations treatment group}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{m1}{First stage sample size}

\item{m}{Pre-planned total sample size}

\item{me}{Extended sample size}

\item{perms}{Maximum number of permutations to use when computing permutation p-values and conditional error rates}
}
\value{
p-value
}
\description{
Adaptive decision rule
}
\details{
Slightly faster than the test that computes p-values
}
\author{
Florian Klinglmueller
}

