% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{compare_tests_onesample}
\alias{compare_tests_onesample}
\title{Compare one-sample test procedures}
\usage{
compare_tests_onesample(B, test_funs, n1, n, rule, rdist, ...)
}
\arguments{
\item{test_funs}{list of test functions (see Details)}

\item{n1}{first stage sample size}

\item{n}{pre-planned total sample size}

\item{rule}{sample size reassessment rule (see Details)}

\item{rdist}{distribution from which to draw observations}

\item{...}{additional arguments to rdist}
}
\value{
let's see
}
\description{
Compare operating characteristics for a number of one-sample tests
}
\details{
\code{test_funs} needs to be a list of functions that take arguments \code{x,n1,n,ne,permutations} and return \code{TRUE} of \code{FALSE} depending on whether the corresponding decision rule rejects the null hypothesis or not.

\code{rule} needs to be a function that takes as argument the vector of first stage observations and return an interger number, which (if larger than \code{n}) will be used as the new total sample size.
}
\author{
Florian Klinglmueller
}

