% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cond_power_rule_norm_ts}
\alias{cond_power_rule_norm_ts}
\title{Conditional power sample size reassessment rule (two-sample z-test)}
\usage{
cond_power_rule_norm_ts(x1, y1, delta = 1, target = 0.9, alpha = 0.025,
  maxN = length(x) * 6, rob_var = T)
}
\arguments{
\item{x1}{First stage observations}

\item{y1}{treatment group obseravtions}

\item{target}{Target power}

\item{alpha}{Significance level}

\item{maxN}{Maximum sample size}

\item{rob_var}{Use robust variance estimate}

\item{m}{Target mean value}
}
\value{
Total sample size required to achieve the target power
}
\description{
Conditional power rule for the two-sample t-test using the function using the normal distribution sample size formula. Reestimates the standard deviation from the first stage and recomputes the sample size such that the power to reject the null meets the target power assuming that the mean (paired treatment difference) is equal to a prespecified value.
}
\details{
In case the robust variance estimate is used we compute the variance as \code{(iqr(c(x1-median(x1),y1-median(y1)))/1.349)^2} this leads to a slight underestimation of the required sample size - for the moment you could increase the target power to correct
}
\author{
Florian Klinglmueller
}

