% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{delta2res}
\alias{delta2res}
\title{Effect size to relative effect size}
\usage{
delta2res(delta, sd = 1)
}
\arguments{
\item{delta}{difference in means}

\item{sd}{common standard deviation}
}
\value{
probability that observations in the treatment sample are larger than in the control sample
}
\description{
Convert difference in means to relative effect size for normally distributed samples
}
\examples{
res <- delta2res(1)
power.u.test(res)
power.z.test(delta=1,sd=1,power=0.84)
power.t.test(delta=1,sd=1,power=0.84)
}
\author{
float
}

