% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iqrp}
\alias{iqrp}
\title{MAD of pooled samples}
\usage{
iqrp(x, y, factor = 1)
}
\arguments{
\item{x}{vector of observations}

\item{y}{vector of observations}

\item{factor}{Scaling factor (deprecated)}
}
\value{
numeric value
}
\description{
Compute the interquartile range pooled over two samples
}
\details{
We subtract the median from each group and then compute the interquartile range from the combination of the (shifted) samples. Finite sample corretion factors have been estimated from a simulation study. Those do not work particularly well to produce nearly unbiased estimates for normally distributed samples but are better than nothing.

The scaling factor does nothing.
}
\author{
Florian Klinglmueller
}

