% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{madp}
\alias{madp}
\title{MAD of pooled samples}
\usage{
madp(x, y, factor = 1)
}
\arguments{
\item{x}{vector of observations}

\item{y}{vector of observations}

\item{factor}{Scaling factor (deprecated)}
}
\value{
numeric value
}
\description{
Compute the median absolute deviation pooled over two samples
}
\details{
We subtract the median from each group and then compute the MAD from the combination of the (shifted) samples. Finite sample corretion factors have been estimated from a simulation study and work well to produce nearly unbiased estimates for normally distributed samples.

The scaling factor does nothing.
}
\author{
Florian Klinglmueller
}

