% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_rules.R
\name{n2lambda}
\alias{n2lambda}
\title{n1 to lambda}
\usage{
n2lambda(n1, target = 0.8)
}
\arguments{
\item{n1}{first stage control group sample size}

\item{target}{target power}
}
\value{
penalty factor for the combined sample size objective
}
\description{
Estimate the penalty factor lambda from n1, in a way to give roughly 80% overall power
}
\details{
Except for first stage sample sizes 5, 15 and 50 this may give totally useless results.
}
\author{
float
}
\seealso{
\code{\link{optimal_power_rule_w_ts}}, \code{\link{combination_power_rule_w_ts}} and  \code{\link{approximate_power_rule_w_ts}} for sample size rules based on a combined objective (maximum conditional power with moderate small sample size)
}

