% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_rules.R
\name{order_probabilities}
\alias{order_probabilities}
\title{Pair-wise order probabilities}
\usage{
order_probabilities(x, y)
}
\arguments{
\item{x}{vector of (control group) observations}

\item{y}{vector of (treatment group) observations}
}
\value{
list of order probabilities
}
\description{
This function computes the three probabilities P(X < Y), P(X < Y and X < Y'), and P(X < Y and X' < Y) [see (2.16), (2.19) and (2.20) on p. 69f of Lehmann (2006)] for two given samples of observations.
}
\author{
float
}

