% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.R
\name{perm_dist}
\alias{perm_dist}
\title{Permutation distribution}
\usage{
perm_dist(x1, x2, g1, g2, stat, B, x3 = NULL, g3 = NULL,
  restricted = TRUE, stratified = TRUE)
}
\arguments{
\item{x1}{First stage data}

\item{x2}{Second stage data}

\item{g1}{(Dummy) first stage treatment assignments}

\item{g2}{(Dummy) second stage treatment assignments (see Details)}

\item{stat}{Function that computes the test statistic (signature \code{function(x,g)})}

\item{B}{Number of permutations}

\item{x3}{(Dummy) third stage data (e.g. sample size increase)}

\item{g3}{(Dummy) third stage treatment assignments (see Details)}

\item{restricted}{Should group sizes be considered fixed}

\item{stratified}{Should permutation be stratified by stage}
}
\value{
numeric vector
}
\description{
Compute the permutation distribution given observations and a test statistic.
}
\details{
First, second and third stage treatment assignments are only passed to define the second stage sample size and the sizes of the corresponding (treatment) groups. If the number of permutations is less than the number of all possible permutations the test observed test statistic will be added to the result such that $B+1$ values will be returned. If the number of requested permutations \code{B} is larger than the number of all possible permutations, only the latter will be used.
}
\author{
Florian Klinglmueller
}

