% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{power.u.test}
\alias{power.u.test}
\title{sample size formula for the (one-sided) Mann-Whitney U test}
\usage{
power.u.test(res, propn = 1/2, sig.level = 0.025, power = 0.8,
  silent = F)
}
\arguments{
\item{res}{relative effect size, i.e. probability that Y_i > X_i}

\item{propn}{proportion of control group samples among total sample}

\item{sig.level}{significance level}

\item{power}{desired power}

\item{silent}{should explanations be suppressed}

\item{n}{total sample size}
}
\value{
sample size (rounded to the next larger integer)
}
\examples{
power.z.test(delta=1,sd=1,power=.84)
power.t.test(delta=1,sd=1,power=.84)
power.u.test(delta2res(delta=1,sd=1),power=.8)
}
\author{
Florian Klinglmueller
}
\references{
Noether, Gottfried E. "Sample size determination for some common nonparametric tests." Journal of the American Statistical Association 82.398 (1987): 645-647.
}

