% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_rules.R
\name{power.w.test}
\alias{power.w.test}
\title{Power and sample size of the Wilcoxon Mann Whitney U test}
\usage{
power.w.test(n = NULL, p1 = NULL, p2 = NULL, p3 = NULL, propn = 1/2,
  sig.level = 0.025, power = 0.9, silent = F)
}
\arguments{
\item{n}{Sample size in the control group}

\item{p1}{probability that X < Y}

\item{p2}{probability that X < Y and X < Y'}

\item{p3}{probability that X < Y and X' < Y}

\item{propn}{proportion of the total number of observations in the control group}

\item{sig.level}{significance level of the test}

\item{power}{desired target power}

\item{silent}{should hints be suppressed}
}
\value{
Either the required number of control group observations to achieve the targe power, or the power of the WMW test under the specified alternative and given sample size.
}
\description{
This function computes the power and sample size according to the asymptotic power formula given in [(2.23) p.71 of Lehmann (2006)]. If \code{is.null(n)} the sample size required to achieve the target power is computed using bisection search.
}
\author{
float
}

