% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{robust_pooled_variance}
\alias{robust_pooled_variance}
\title{Robust pooled variance estimate}
\usage{
robust_pooled_variance(x, y, type = c("qn", "sn", "iqr", "mad"),
  factor = NULL)
}
\arguments{
\item{x}{control group observations}

\item{y}{treatment group observations}

\item{type}{what estimator to use}

\item{factor}{multiplication constant}
}
\value{
robust variance estimate
}
\author{
float
}

