% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{snp}
\alias{snp}
\title{Robust scale estimate Qn pooled sample}
\usage{
snp(x, y, factor = 1.1926)
}
\arguments{
\item{x}{vector of observations}

\item{y}{vector of observations}

\item{factor}{Scaling factor}
}
\value{
numeric value
}
\description{
Compute the robust scale measure Sn pooled over two samples.
}
\details{
The default scaling factor is chosen such that the estimate is approximately consistent for normal observations. For different reference distributions other factors may be more apropriate. Note however, that we have hardcoded the bias correction constants applied to small samples - these may differ as well for other reference distributions. Changing the factor may be futile.

Note that we use a computationally suboptimal implementation that uses \code{O(n^2)} operations. If time allows we may implement the more efficient method given in "Croux and Rousseeuw (1992)" which requires only \code{O(n*log(n))} operations.
}
\author{
Florian Klinglmueller
}

