% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_sample_os}
\alias{split_sample_os}
\title{Split single group sample}
\usage{
split_sample_os(x, n1, n, ne = n)
}
\arguments{
\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}
}
\value{
list with stagewise observations and stagewise sign-indicators
}
\description{
Split sample of a single group experiment into stages
}
\author{
Florian Klinglmueller
}

