% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_rules.R
\name{SSR_WMW}
\alias{SSR_WMW}
\alias{approximate_power_rule_w_ts}
\alias{combination_power_rule_w_ts}
\alias{cond_power_rule_w_ts}
\alias{optimal_power_rule_w_ts}
\alias{predictive_power_rule_w_ts}
\title{Sample size reassessment rule for rank based randomization tests}
\usage{
cond_power_rule_w_ts(x1, y1, z2, pp, target = 0.9, alpha = 0.025,
  maxN = length(x1) * 6, propn = 1/2)

predictive_power_rule_w_ts(x1, y1, m = 2 * length(x1), target = 0.9,
  alpha = 0.025, maxN = length(x1) * 6, propn = 1/2)

approximate_power_rule_w_ts(x1, y1, m = 2 * length(x1), pp, lambda = 1e-04,
  alpha = 0.025, maxN = length(x1) * 6, propn = 1/2)

combination_power_rule_w_ts(x1, y1, m = 2 * length(x1), pp, lambda = 1e-04,
  alpha = 0.025, maxN = length(x1) * 6, propn = 1/2)

optimal_power_rule_w_ts(x1, y1, z2, pp, lambda = 1e-04, alpha = 0.025,
  maxN = length(x1) * 6, propn = 1/2)
}
\arguments{
\item{x1}{first stage (control-group) sample}

\item{y1}{first stage (control-group) sample}

\item{z2}{second stage blinded combined sample}

\item{pp}{list with prespecified order probabilities p1, p2, p3 (see \code{\link{order_probabilities}} and \code{\link{delta2res}} for details)}

\item{target}{desired target power}

\item{alpha}{significance level of the preplanned test}

\item{maxN}{maximum overall (control-group) sample size}

\item{propn}{proportion of the total number of observations in the control group}

\item{m}{preplanned overall (control-goup) sample size}

\item{lambda}{penalty factor for additional sample in the combined objective}
}
\value{
reassessed overall sample size
}
\description{
Re-computes the sample size based on first stage observations and assuming that a WMW test is performed at the end of the trial. It estimates the overall (across all stages) number of control-group observations required to achieve the target power. The returned value may be smaller than the preplanned sample size, even the first stage sample size - so any minimum sample size restrictions need to be enforced outside of the function. The maximum (overall control-group) sample size however is enforced internally.
}
\details{
The function \code{cond_power_rule_w_ts} recomputes the sample size using the sample size formula of \code{\link{power.w.test}} to achieve the target power \code{target} assuming the order probabilities \code{pp} with a second stage WMW test at the level of the conditional error rate of the preplanned test.

The function \code{predictive_power_rule_w_ts} recomputes the order probabilities based on the first stage observations and reestimates the sample size required to achieve the target power using the sample size formula of \code{\link{power.w.test}}.

The functions \code{approximate_power_rule_w_ts}, \code{combination_power_rule_w_ts} and \code{optimal_power_rule_w_ts} compute the sample size that maximizes the combined objective \code{CP(x1,y1) - lambda*nA} defined in [Jennison and Turnbull (2015)]. \code{approximate_power_rule_w} estimates the conditional power using the power formula of \code{\link{power.w.test}} with weighted averages of the prespecified order proabibilities \code{pp} and the order probabilities estimated from the first stage obseravtions. \code{comb_power_rule_w_ts} estimates the conditional power using the power formula of \code{\link{power.w.test}} of a second stage WMW with the prespecified order probabilities at the level of the conditional error of the inverse normal combination WMW test, finally \link{optimal_power_rule_w_ts} estimates the conditional power in the same way however using the conditional error rate of the preplanned randomization test - which requires knowledge of the blinded second stage observations.
}
\section{Functions}{
\itemize{
\item \code{cond_power_rule_w_ts}: Conditional power rule for the WMW test

\item \code{predictive_power_rule_w_ts}: Predicitive power rule for the WMW test

\item \code{approximate_power_rule_w_ts}: Efficient power rule for the WMW test using approximate conditional order probabilities

\item \code{combination_power_rule_w_ts}: Efficient power rule for the WMW test using the combination test conditional power

\item \code{optimal_power_rule_w_ts}: Efficient power rule for the WMW test using the conditional power of a second stage WMW test
}}
\author{
float
}
\references{
Lehmann, Erich Leo, and H. J. D'abrera. Nonparametrics: statistical methods based on ranks. Springer, 2006.

            Jennison, Christopher, and Bruce W. Turnbull. "Adaptive sample size modification in clinical trials: start small then ask for more?." Statistics in medicine 34.29 (2015): 3793-3810.

            Bauer, Peter, and Franz Koenig. "The reassessment of trial perspectives from interim data—a critical view." Statistics in medicine 25.1 (2006): 23-36.
}
\seealso{
\code{\link{order_probabilities}} that estimates order probabilities from observations, \code{\link{delta2res}} that computes order probabilisities for normally distributed observations
}

