% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{t2p}
\alias{t2p}
\title{Permutation p-value}
\usage{
t2p(t, dist, type = c("non-randomized", "midp", "randomized",
  "davison_hinkley"))
}
\arguments{
\item{t}{observed test statistic}

\item{dist}{permutation distribution}

\item{type}{type of p-value to compute}
}
\value{
p-value
}
\description{
Permutation test p-value
}
\details{
Type "non-randomized" will give the standard p-value, which is define by the proportion of sample permutations with larger or equal test statistic than the observed one. "midp" will give the mid-p value which is the proportion of sample permutations with larger test statistic plus one half the proportion of permutations with equal test statistic. "randomized" will return the proportion of permutations with larger test statistic plus a random number that is uniformly distributed between zero and the proportion of permutations with test statistic equal to the observed.
}
\author{
Florian Klinglmueller
}

