% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{trimmings}
\alias{trimmings}
\title{trimmings}
\usage{
trimmings(g1, g2, restricted, alpha = 0.025)
}
\arguments{
\item{g1}{first stage treatment group assignments}

\item{g2}{second stage treatment group assignmens}

\item{restricted}{should permutation be restricted by group size stratified by stage}

\item{alpha}{significance level}
}
\value{
difference between size and significance level
}
\description{
Difference between significance level and actual size of the permutation test, due to discreteness
}
\author{
Florian Klinglmueller
}

