% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaperm-package.R
\docType{package}
\name{adaperm-package}
\alias{adaperm-package}
\title{Permutation tests for adaptive designs

Permutation and re-randomization tests for adaptive designs.

Provides functions to perform permutation tests.}

