% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_permtest.R
\name{adaperm_DR}
\alias{adaperm_DR}
\title{Permutation based decision rule for adaptive designs}
\usage{
adaperm_DR(x, g = NULL, n1, n, m1 = n1, m = n, test_statistic,
  alpha = 0.025, cer_type = .cer_types, atest_type = .atest_types,
  permutations = 10000)
}
\arguments{
\item{x}{Observations}

\item{g}{Treatment assignments (if \code{NULL}) a one-sample test will be performed}

\item{n1}{First stage control-group sample size}

\item{n}{Pre-planned overall sample size}

\item{m1}{First stage treatment-group sample size (will be ignored if \code{g} is \code{NULL})}

\item{m}{Pre-planned overall treatment-group sample size (will be ignored if \code{g} is \code{NULL})}

\item{test_statistic}{Test statistic}

\item{alpha}{Significance level}

\item{cer_type}{what type of conditional error rate function should be used (see \code{\link{permutation_cer}}) for detailcer_type}

\item{atest_type}{if 'CER' compute only conditional error rate, else type of adaptive test should be performed (see \code{\link{perm_test}} for details)}

\item{permutations}{Number of permutations to use}
}
\value{
Decision \code{TRUE} if null hypothesis is rejected
}
\description{
User friendly wrapper to \code{\link{adaptive_permdr}}
}
\author{
Florian Klinglmueller
}

