% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{robust_pooled_variance}
\alias{robust_pooled_variance}
\title{Robust pooled variance estimate}
\usage{
robust_pooled_variance(x, y)
}
\arguments{
\item{x}{control group observations}

\item{y}{treatment group observations}
}
\value{
robust variance estimate
}
\description{
Robust pooled variance estimate
}
\author{
float
}

